+++ packages/frontend/src/features/export/index.ts
@@ -1,0 +1,1 @@
+export * from './components/ExportToGoogleSheet';
